Add the SiteSendEcho interface and move events API to its own package
