+++ src/main/kotlin/markdown/echo/events/EventDsl.kt
@@ -1,4 +1,4 @@
-package markdown.echo
+package markdown.echo.events
 
 /**
  * A [DslMarker] for APIs that are part of the event DSL of the Echo API.
