+++ src/main/kotlin/markdown/echo/events/EventScope.kt
@@ -1,4 +1,4 @@
-package markdown.echo
+package markdown.echo.events
 
 import markdown.echo.causal.EventIdentifier
